Implementation of A and K operators
The model code was running on an algebraic expression for the FDM based
on the operators for homogeneous media. The algebraic expression was
changed to incorporate the A and K operators, so that these operators
can be altered for heterogeneous media while still maintaining the
model code.

For clarity, the parameter “value_du_dxx” was changed to
“value_mu_du_dxx”, since the expression now contains the operator K,
which already incorporates the multiplication with mu.

Updating of the displacement was simplified so it will work when
implementing further derivatives.